requirements.txt
Django==4.1.3
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
python-dotenv==0.21.0
gunicorn==20.1.0
whitenoise==6.2.0
dj-database-url==1.0.0
psycopg2-binary==2.9.5
python-decouple==3.6